+++ lib/app/data/models/component_model.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:habitat_ft_user/app/utils/config/colors.dart';
 import 'package:habitat_ft_user/app/utils/enums.dart';
-import 'package:habitat_ft_user/app/utils/mapper.dart';
 
+// TODO refactor de todo
 class Component {
   // static const COLLECTION_NAME = 'components';
   String title;
@@ -17,7 +18,7 @@ class Component {
   Component.fromJson(Map<String, dynamic> json) {
     title = json['title'];
     url = json['url'];
-    mediaType = Mapper.intToMediaType(json['media_type']);
+    mediaType = Component.intToMediaType(json['media_type']);
   }
 
   Map<String, dynamic> toJson() {
@@ -27,7 +28,57 @@ class Component {
     return data;
   }
 
-  Color get color => Mapper.mediaTypeToColor(mediaType);
+  Color get color => Component.mediaTypeToColor(mediaType);
 
-  IconData get icon => Mapper.mediaTypeToIconData(mediaType);
+  IconData get icon => Component.mediaTypeToIconData(mediaType);
+
+  static Component byJson(json) => Component.fromJson(json);
+
+  static ComponentMediaType intToMediaType(int value) {
+    switch (value) {
+      case 0:
+        return ComponentMediaType.video;
+      case 1:
+        return ComponentMediaType.image;
+      case 2:
+        return ComponentMediaType.file;
+      case 3:
+        return ComponentMediaType.audio;
+      default:
+        print('No se encontro un indice para mapear un MediaType');
+        return ComponentMediaType.audio;
+    }
+  }
+
+  static Color mediaTypeToColor(ComponentMediaType mediaType) {
+    switch (mediaType) {
+      case ComponentMediaType.video:
+        return CustomerColors.COMPLEMENTO_1;
+      case ComponentMediaType.image:
+        return CustomerColors.COMPLEMENTO_2;
+      case ComponentMediaType.file:
+        return CustomerColors.COMPLEMENTO_3;
+      case ComponentMediaType.audio:
+        return CustomerColors.COMPLEMENTO_4;
+      default:
+        print('No se encontro un indice para mapear un MediaType-Color');
+        return CustomerColors.CELESTE_HABITAT;
+    }
+  }
+
+  static IconData mediaTypeToIconData(ComponentMediaType mediaType) {
+    switch (mediaType) {
+      case ComponentMediaType.video:
+        return Icons.videocam;
+      case ComponentMediaType.image:
+        return Icons.image;
+      case ComponentMediaType.file:
+        return Icons.file_present;
+      case ComponentMediaType.audio:
+        return Icons.volume_up;
+      default:
+        print('No se encontro un indice para mapear un MediaType-IconData');
+        return Icons.error;
+    }
+  }
 }
