+++ lib/app/modules/home/widgets/customer_drawer/widgets/customer_drawer.dart/sign_out_tile/widgets/sign_out_alert_dialog.dart
@@ -8,12 +8,10 @@ class SignOutAlertDialog extends StatelessWidget {
   const SignOutAlertDialog();
 
   @override
-  Widget build(BuildContext context) {
-    return CustomerAlertDialog(
-      children: [
-        const SignOutTitle(),
-        const SignOutConfirmAlertDialogButton(),
-      ],
-    );
-  }
+  Widget build(_) => CustomerAlertDialog(
+        children: [
+          const SignOutTitle(),
+          const SignOutConfirmAlertDialogButton(),
+        ],
+      );
 }
