+++ lib/app/modules/home/widgets/customer_drawer/widgets/customer_drawer.dart/sign_out_tile/widgets/sign_out_confirm_alert_dialog_buttons.dart
@@ -7,7 +7,7 @@ class SignOutConfirmAlertDialogButton extends GetView<AuthRepository> {
   const SignOutConfirmAlertDialogButton();
 
   @override
-  Widget build(BuildContext context) => CustomerConfirmAlertButton(
+  Widget build(_) => CustomerConfirmAlertButton(
         onCancelPressed: Get.back,
         onConfirmPressed: controller.signOut,
       );
