+++ lib/app/modules/home/widgets/customer_drawer/widgets/customer_drawer.dart/sign_out_tile/widgets/sign_out_title.dart
@@ -8,5 +8,5 @@ class SignOutTitle extends StatelessWidget {
   TextStyle get style => CustomerStyles.TITULO_3_CELESTE;
 
   @override
-  Widget build(BuildContext context) => Text(title, style: style);
+  Widget build(_) => Text(title, style: style);
 }
